+++ src/crosschain/toucanRelay/ProposalIdCodec.sol
@@ -1,6 +1,6 @@
 pragma solidity ^0.8.0;
 
-/// at the moment this is the same as abi.encodePacked but we're thinking through the problem
+/// at the moment this is the same as abi.encodePacked but we're thinking through the problem in fixed size chunks
 library ProposalIdCodec {
     // we have 32 bits of unused space
     function encode(
@@ -10,18 +10,18 @@ library ProposalIdCodec {
     ) internal pure returns (uint256 proposalId) {
         uint256 addr = uint256(uint160(_plugin));
         return
-            (addr << 96) |
-            (uint256(_proposalStartTimestamp) << 64) |
-            ((uint256(_proposalEndTimestamp)) << 32);
-        // 32 bits of unused space
+            (addr << 64) |
+            (uint256(_proposalStartTimestamp) << 32) |
+            uint256(_proposalEndTimestamp);
     }
 
     function decode(
         uint256 _proposalId
     ) internal pure returns (address plugin, uint32 startTimestamp, uint32 endtimestamp) {
         // shift out the redundant bits then cast to the correct type
-        plugin = address(uint160(_proposalId >> 96));
-        startTimestamp = uint32(_proposalId >> 64);
-        endtimestamp = uint32(_proposalId >> 32);
+        plugin = address(uint160(_proposalId >> 64));
+        startTimestamp = uint32(_proposalId >> 32);
+        endtimestamp = uint32(_proposalId);
     }
 }
+// Let's explore a delay