+++ src/voting/IMajorityVoting.sol
@@ -7,7 +7,18 @@ pragma solidity ^0.8.8;
 /// @notice The interface of majority voting plugin.
 /// @custom:security-contact sirt@aragon.org
 interface IMajorityVoting {
+    /// @notice A container for the proposal vote tally.
+    /// @param abstain The number of abstain votes casted.
+    /// @param yes The number of yes votes casted.
+    /// @param no The number of no votes casted.
+    struct Tally {
+        uint256 abstain;
+        uint256 yes;
+        uint256 no;
+    }
+
     /// @notice Vote options that a voter can chose from.
+    /// @dev Voters can now choose to split their votes across multiple options.
     /// @param None The default option state of a voter indicating the absence from the vote.
     /// This option neither influences support nor participation.
     /// @param Abstain This option does not influence the support but counts towards participation.
@@ -20,17 +31,15 @@ interface IMajorityVoting {
         No
     }
 
-    /// @notice Emitted when a vote is cast by a voter.
+    /// @notice Emitted when votes are cast by a voter.
     /// @param proposalId The ID of the proposal.
     /// @param voter The voter casting the vote.
-    /// @param voteOption The casted vote option.
-    /// @param votingPower The voting power behind this vote.
-    event VoteCast(
-        uint256 indexed proposalId,
-        address indexed voter,
-        VoteOption voteOption,
-        uint256 votingPower
-    );
+    /// @param voteOptions The vote options casted by the voter.
+    event VoteCast(uint256 indexed proposalId, address indexed voter, Tally voteOptions);
+
+    /// @notice DEPRECATED: Emitted when a vote is cast by a voter.
+    /// @dev Voters can now choose to split their votes across multiple options.
+    event VoteCast(uint256 indexed, address indexed, VoteOption, uint256);
 
     /// @notice Returns the support threshold parameter stored in the voting settings.
     /// @return The support threshold parameter.
@@ -68,9 +77,17 @@ interface IMajorityVoting {
     /// - the voter doesn't have voting powers.
     /// @param _proposalId The proposal Id.
     /// @param _account The account address to be checked.
-    /// @param  _voteOption Whether the voter abstains, supports or opposes the proposal.
+    /// @param  _voteOptions Degree to which the voter abstains, supports or opposes the proposal.
     /// @return Returns true if the account is allowed to vote.
     /// @dev The function assumes the queried proposal exists.
+    function canVote(
+        uint256 _proposalId,
+        address _account,
+        Tally memory _voteOptions
+    ) external view returns (bool);
+
+    /// @notice overload of the above function that allows for a legacy vote option
+    /// @param _voteOption The vote option to be checked.
     function canVote(
         uint256 _proposalId,
         address _account,
@@ -82,25 +99,30 @@ interface IMajorityVoting {
     /// @return True if the proposal can be executed, false otherwise.
     function canExecute(uint256 _proposalId) external view returns (bool);
 
-    /// @notice Votes for a vote option and, optionally, executes the proposal.
-    /// @dev `_voteOption`, 1 -> abstain, 2 -> yes, 3 -> no
+    /// @notice Votes for a set of vote options and, optionally, executes the proposal.
     /// @param _proposalId The ID of the proposal.
-    /// @param _voteOption The chosen vote option.
+    /// @param _voteOptions The chosen vote options.
     /// @param _tryEarlyExecution If `true`,  early execution is tried after the vote cast.
     /// The call does not revert if early execution is not possible.
+    function vote(uint256 _proposalId, Tally memory _voteOptions, bool _tryEarlyExecution) external;
+
+    /// @notice Supports voting for a single vote options with full voting power using legacy voting
     function vote(uint256 _proposalId, VoteOption _voteOption, bool _tryEarlyExecution) external;
 
     /// @notice Executes a proposal.
     /// @param _proposalId The ID of the proposal to be executed.
     function execute(uint256 _proposalId) external;
 
-    /// @notice Returns whether the account has voted for the proposal.
-    /// Note, that this does not check if the account has voting power.
+    /// @notice Returns the latest votes of the account for the proposal.
     /// @param _proposalId The ID of the proposal.
     /// @param _account The account address to be checked.
-    /// @return The vote option cast by a voter for a certain proposal.
-    function getVoteOption(
+    /// @return The vote options cast by a voter for the proposal.
+    function getVoteOptions(
         uint256 _proposalId,
         address _account
-    ) external view returns (VoteOption);
+    ) external view returns (Tally memory);
+
+    /// @notice DEPRECATED: PLEASE USE `getVoteOptions` INSTEAD.
+    /// @dev supports legacy queries for backwards compatibility and historical state.
+    function getVoteOption(uint256, address) external view returns (VoteOption);
 }